refactor: creating separeted file to fetch functions
